+++ packages/antd-management-fast-design-playground/CHANGELOG.md
@@ -1,5 +1,16 @@
 # antd-management-fast-design-playground
 
+## 1.0.242
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-react@1.0.161
+  - antd-management-fast-component@2.7.86
+  - antd-management-fast-formily@1.0.181
+  - antd-management-fast-common@2.6.189
+
 ## 1.0.241
 
 ### Patch Changes